refactor: proactive.py 按类边界拆分（批6-1）
- ProactiveMemory → proactive_memory.py
- ProactiveLimiter + effective_proactive_config 及常量 → proactive_limiter.py
- proactive.py 保留纯函数区 + ProactiveScreenWatcher + re-export 兼容层
- 逐字节搬移零逻辑改动，import 单向无环

结构优化批6-1（STRUCT_PLAN_FINAL.md）
